+++ cs/pfg/mipago/browser/confirmation.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
+from Acquisition import aq_parent
 from cs.pfg.mipago.config import ANNOTATION_KEY
+from cs.pfg.mipago.config import PAYMENT_STATUS_ERROR_IN_MIPAGO
 from cs.pfg.mipago.config import PAYMENT_STATUS_PAYED
 from cs.pfg.mipago.config import PAYMENT_STATUS_SENT_TO_MIPAGO
-from cs.pfg.mipago.config import PAYMENT_STATUS_ERROR_IN_MIPAGO
 from cs.pfg.mipago.config import PAYMENT_STATUS_USER_IN_MIPAGO
 from plone.protect.interfaces import IDisableCSRFProtection
+from Products.CMFPlone.utils import safe_hasattr
 from Products.Five.browser import BrowserView
 from zope.annotation.interfaces import IAnnotations
 from zope.interface import alsoProvides
@@ -13,14 +15,26 @@ import xml.etree.ElementTree as ET
 
 
 class PaymentConfirmation(BrowserView):
+
+    def get_annotation_context(self):
+        parent = aq_parent(self.context)
+        if safe_hasattr(parent, 'isTranslation') and \
+               parent.isTranslation() and not parent.isCanonical():
+            # look in the canonical version to see if there is
+            # a matching (by id) save-data adapter.
+            # If so, call its onSuccess method
+            cf = parent.getCanonical()
+            target = cf.get(self.context.getId())
+            if target is not None and target.meta_type == 'MiPagoAdapter':
+                return target
+
+        return self.context
+
+
     def __call__(self):
         alsoProvides(self.request, IDisableCSRFProtection)
-        try:
-            canonical = self.context.getCanonical()
-        except:
-            canonical = self.context
-
-        adapted = IAnnotations(canonical)
+        annotation_context = self.get_annotation_context()
+        adapted = IAnnotations(annotation_context)
         payment_code = self.extract_payment_code()
         payment_status = self.extract_payment_status()
         payments = adapted.get(ANNOTATION_KEY, {})